refactor: 코드리뷰 정리
