+++ apps/client/src/pages/time-table/hooks/use-festival-delete.ts
