+++ apps/client/src/pages/time-table/page/add-festival.tsx
@@ -1,11 +1,11 @@
 import { useNavigate } from 'react-router-dom';
 import { Button, FestivalCard, Header } from '@confeti/design-system';
-import * as styles from './add-festival.css';
+import { routePath } from '@shared/constants/path';
+import { useInfiniteScroll } from '@shared/utils/use-infinite-scroll';
 import useFestivalSelection from '../hooks/use-festival-selection';
 import { useGetFestivalToAdd } from '../hooks/use-get-festival-to-add';
 import { useAddTimeTableFestival } from '@pages/time-table/hooks/use-timetable-festival-mutation';
-import { useInfiniteScroll } from '@shared/utils/use-infinite-scroll';
-import { routePath } from '@shared/constants/path';
+import * as styles from './add-festival.css';
 
 const MAX_SELECTIONS = 3;
